Build towards proper exeption safety guarantees
We are required to preserve all already inserted elements whenever possible.

Currently we would rewind everything that happened during one of operations on uninitialized memory.

This allowed to handle keeping track of size changes separately.

However, this is both error prone and also not possible with the basic exception guarantees.